requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.9.3
astroid==3.2.4
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dbus-python==1.2.16
dill==0.3.9
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==1.12
fonttools==4.55.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.6.0
icecream==2.1.3
idna==2.8
imageio==2.35.1
imgaug==0.4.0
importlib-metadata==4.11.4
importlib-resources==6.4.5
iniconfig==2.0.0
invoke==2.2.0
isort==5.13.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy-loader==0.4
libclang==14.0.1
mando==0.7.1
Markdown==3.3.7
matplotlib==3.7.5
mccabe==0.7.0
networkx==3.1
numpy==1.22.4
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.26.1
pycodestyle==2.12.1
pygments==2.18.0
PyGObject==3.36.0
pylint==3.2.7
pyparsing==3.0.9
pytest==8.3.3
python-apt==2.0.0+ubuntu0.20.4.7
python-box==7.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
radon==6.0.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six==1.14.0
smmap==5.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2023.7.10
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typing-extensions==4.2.0
tzdata==2024.2
urllib3==1.25.8
visual-logging==1.0
Werkzeug==2.1.2
wrapt==1.14.1
xenon==0.9.3
xmltodict==0.14.2
zipp==3.8.0
